// packages/extension-contract/src/types.d.ts
export type PiariumContextValue = string | number | boolean;
export type PiariumContextExpressionV1 = {
    op: "defined";
    key: string;
} | {
    op: "equals";
    key: string;
    value: PiariumContextValue;
} | {
    op: "not";
    expression: PiariumContextExpressionV1;
} | {
    op: "all" | "any";
    expressions: PiariumContextExpressionV1[];
};